+++ eventstore/repository/esdb/connect.go
@@ -2,7 +2,6 @@ package esdb
 
 import (
 	"github.com/EventStore/EventStore-Client-Go/v3/esdb"
-	//"github.com/calogxro/qaservice/config"
 )
 
 func connect(URI string) (*esdb.Client, error) {