+++ projection/cmd/main.go
@@ -1,26 +1,26 @@
 package main
 
 import (
-	"github.com/calogxro/qaservice/domain"
-	esdbgateway "github.com/calogxro/qaservice/projection/gateway/esdb"
-	httphandler "github.com/calogxro/qaservice/projection/handler/http"
-	"github.com/calogxro/qaservice/projection/repository/mongodb"
+	"fmt"
+
+	"github.com/calogxro/qaservice/config"
+	"github.com/calogxro/qaservice/projection/factory"
 	"github.com/calogxro/qaservice/projection/service/projection"
-	"github.com/calogxro/qaservice/projection/service/projector"
+
+	//"github.com/calogxro/qaservice/domain"
+	//esdbgateway "github.com/calogxro/qaservice/projection/gateway/esdb"
+
+	"github.com/calogxro/qaservice/projection/repository/mongodb"
+
+	//"github.com/calogxro/qaservice/projection/service/projector"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
-	readRepository := mongodb.New()
-	p := projector.New(readRepository)
+	fmt.Println(config.MongoDB)
 
-	esgw := esdbgateway.New()
-	go esgw.Subscribe(func(event *domain.Event) {
-		p.Project(event)
-	})
-
-	service := projection.New(readRepository)
-	router := httphandler.MakeHandler(service, gin.New())
-	router.Run(":8082")
+	var repo projection.ReadRepository = mongodb.New()
+	router := factory.MakeServer(repo, gin.New())
+	router.Run(":8081")
 	//router.Run(config.IP_PORT)
 }